chore: Rename utils folder to util
